added java-saml-core module, left servlet dependencies in java-saml
updated SamlResponse to take in the requestURL and SAMLResponse
parameter directly. the new module does not depend on
javax.servlet.*, meaning it can more easily be reused together with
different server stacks.

improved Util.getFileAsString() to work for classpath resources that are
in JARs as well (previously only on the filesystem worked).
