--- core/.gitignore
@@ -1,6 +1,0 @@
-target
-.idea
-*.iml
-.classpath
-.settings/
-.project
