--- core/src/test/java/com/onelogin/saml2/test/ConsoleTestRunner.java
@@ -1,14 +1,0 @@
-package com.onelogin.saml2.test;
-
-import org.junit.runner.JUnitCore;
-import org.junit.runner.Result;
-import org.junit.runner.notification.Failure;
-
-public class ConsoleTestRunner {
-	public static void main(String[] args) {
-		Result result = JUnitCore.runClasses(AllTestsSuite.class);
-		for (Failure failure : result.getFailures()) {
-			System.out.println(failure.toString());
-		}
-	}
-}
