+++ pom.xml
@@ -6,7 +6,7 @@
 	<version>2.0.0-SNAPSHOT</version>
 	<packaging>pom</packaging>
 
-	<name>OneLogin java-saml Toolkit</name>
+	<name>OneLogin java-saml Toolkit Project</name>
 	<description>A Java SAML toolkit by OneLogin</description>
 	<url>https://github.com/onelogin/java-saml</url>
 
@@ -20,6 +20,7 @@
 
 	<modules>
 		<module>core</module>
+		<module>toolkit</module>
 		<module>samples</module>
 	</modules>
