+++ toolkit/src/main/java/com/onelogin/saml2/Auth.java
@@ -26,6 +26,7 @@ import com.onelogin.saml2.exception.XMLEntityException;
 import com.onelogin.saml2.exception.SettingsException;
 import com.onelogin.saml2.logout.LogoutRequest;
 import com.onelogin.saml2.logout.LogoutResponse;
+import com.onelogin.saml2.servlet.ServletUtils;
 import com.onelogin.saml2.settings.Saml2Settings;
 import com.onelogin.saml2.settings.SettingsBuilder;
 import com.onelogin.saml2.util.Constants;
@@ -220,7 +221,7 @@ public class Auth {
 
 		String relayState;
 		if (returnTo == null) {
-			relayState = Util.getSelfRoutedURLNoQuery(request);
+			relayState = ServletUtils.getSelfRoutedURLNoQuery(request);
 		} else {
 			relayState = returnTo;
 		}
@@ -237,7 +238,7 @@ public class Auth {
 		String ssoUrl = getSSOurl();
 
 		LOGGER.debug("AuthNRequest sent to " + ssoUrl + " --> " + samlRequest);
-		Util.sendRedirect(response, ssoUrl, parameters);
+		ServletUtils.sendRedirect(response, ssoUrl, parameters);
 		lastRequestId = authnRequest.getId();
 	}
 
@@ -284,7 +285,7 @@ public class Auth {
 
 		String relayState;
 		if (returnTo == null || returnTo.isEmpty()) {
-			relayState = Util.getSelfRoutedURLNoQuery(request);
+			relayState = ServletUtils.getSelfRoutedURLNoQuery(request);
 		} else {
 			relayState = returnTo;
 		}
@@ -301,7 +302,7 @@ public class Auth {
 
 		String sloUrl = getSLOurl();
 		LOGGER.debug("Logout request sent to " + sloUrl + " --> " + samlLogoutRequest);
-		Util.sendRedirect(response, sloUrl, parameters);
+		ServletUtils.sendRedirect(response, sloUrl, parameters);
 		lastRequestId = logoutRequest.getId();
 	}
 
@@ -460,7 +461,7 @@ public class Auth {
 
 				String sloUrl = getSLOurl();
 				LOGGER.debug("Logout response sent to " + sloUrl + " --> " + samlLogoutResponse);
-				Util.sendRedirect(response, sloUrl, parameters);				
+				ServletUtils.sendRedirect(response, sloUrl, parameters);
 			}
 		} else {
 			errors.add("invalid_binding");
