+++ toolkit/src/main/java/com/onelogin/saml2/logout/LogoutResponse.java
